# openapi/v1.yaml
...
components:
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  title: WaddleAI Management API
  version: '1'
openapi: 3.1.0
paths:
...
  /api/v1/routing/policies/{organization_id}:
    delete:
      description: ''
      operationId: delete_delete_policy
      parameters:
      - in: path
        name: organization_id
        required: true
        schema:
          type: number
      responses:
        '200': *id001
      summary: Delete an org's routing policy row (admin only) -- resets it to engine
        defaults.
    get:
      description: ''
      operationId: get_get_policy
      parameters:
      - in: path
        name: organization_id
        required: true
        schema:
          type: number
      responses:
        '200': *id001
      summary: Get an org's routing policy, or engine defaults if no row exists yet.
    put:
      description: ''
      operationId: put_upsert_policy
      parameters:
      - in: path
        name: organization_id
        required: true
        schema:
          type: number
      responses:
        '200': *id001
      summary: Create or update an org's routing policy (upsert on organization_id).
...
  /api/v1/security-policies/bypass-grants/{grant_id}:
    delete:
      description: ''
      operationId: delete_revoke_bypass_grant
      parameters:
      - in: path
        name: grant_id
        required: true
        schema:
          type: number
      responses:
        '200': *id001
      summary: Revoke (delete) a bypass grant, org-scoped for non-admin roles.
...
  /api/v1/security-policies/{policy_id}:
    delete:
      description: ''
      operationId: delete_security_policy
      parameters:
      - in: path
        name: policy_id
        required: true
        schema:
          type: number
      responses:
        '200': *id001
      summary: Delete a security policy by ID.
    put:
      description: ''
      operationId: put_update_policy
      parameters:
      - in: path
        name: policy_id
        required: true
        schema:
          type: number
      responses:
        '200': *id001
      summary: Update selected fields of an existing security policy.
...